+++ examples/gauge/main.py
@@ -29,8 +29,10 @@ nh.nexInit()
 # create a gauge instance
 z0 = NexGauge(nh, 0, 1, "z0")
 
-# new value of gauge
+# new values of gauge
 gauge_value = randint(10, 350)
+background_color_value = 63488  # red
+font_color_value = 31           # blue
 
 # request the value of gauge "z0"
 print('Requesting gauge "{}" value ...'.format(z0.name))
@@ -55,6 +57,61 @@ print()
 
 # sanity check
 if response != gauge_value:
+    print('WARNING: GET value did not match SET value')
+
+# request the background color of gauge "z0"
+print('Requesting background color of gauge "{}" ...'.format(z0.name))
+response = z0.Get_background_color_bco()
+print('Background color of gauge "{}" is: "{}"'.format(z0.name, response))
+print()
+
+time.sleep(1)
+
+# modify the background color of gauge "z0" to "red"
+# search for RGB565 Colors. Red is "63488" at 65k colors
+print('Set background color of gauge "{}" to "{}"'.
+      format(z0.name, background_color_value))
+z0.Set_background_color_bco(background_color_value)
+print()
+
+time.sleep(1)
+
+# request the background color of gauge "z0" again
+print('Requesting background color of gauge "{}" ...'.format(z0.name))
+response = z0.Get_background_color_bco()
+print('Background color of gauge "{}" is: "{}"'.format(z0.name, response))
+print()
+
+# sanity check
+if response != background_color_value:
+    print('WARNING: GET value did not match SET value')
+
+time.sleep(1)
+
+# request the font color of gauge "z0"
+print('Requesting font color of gauge "{}" ...'.format(z0.name))
+response = z0.Get_font_color_pco()
+print('Font color of gauge "{}" is: "{}"'.format(z0.name, response))
+print()
+
+time.sleep(1)
+
+# modify the font color of gauge "z0" to "blue"
+# search for RGB565 Colors. Blue is "31" at 65k colors
+print('Set font color of gauge "{}" to "{}"'.format(z0.name, font_color_value))
+z0.Set_font_color_pco(font_color_value)
+print()
+
+time.sleep(1)
+
+# request the font color of gauge "z0" again
+print('Requesting font color of gauge "{}" ...'.format(z0.name))
+response = z0.Get_font_color_pco()
+print('Font color of gauge "{}" is: "{}"'.format(z0.name, response))
+print()
+
+# sanity check
+if response != font_color_value:
     print('WARNING: GET value did not match SET value')
 
 print('Returning to REPL in 5 seconds')